+++ cliquet_fxa/tests/test_authentication.py
@@ -44,7 +44,7 @@ class FxAOAuthAuthenticationPolicyTest(unittest.TestCase):
         self.request.registry.cache = self.backend
         settings = DEFAULT_SETTINGS.copy()
         settings['fxa-oauth.cache_ttl_seconds'] = '0.01'
-        settings['fxa-oauth.mandatory_scope'] = 'mandatory profile'
+        settings['fxa-oauth.required_scope'] = 'mandatory profile'
         self.request.registry.settings = settings
         self.request.headers['Authorization'] = 'Bearer foo'
         self.profile_data = {